CCSS loader data parser
