--- lib/tasks/ccls.rake
@@ -1,6 +1,0 @@
-namespace :ccls do
-  desc 'Load CCLS standards'
-  task :load => :environment do
-    CclsTasks.load
-  end
-end
