+++ lib/tasks/ccss.rake
@@ -1,0 +1,6 @@
+namespace :ccss do
+  desc 'Load CCSS standards'
+  task :load => :environment do
+    CcssTasks.load
+  end
+end
