+++ scripts/fetch_standards_json.sh
@@ -1,7 +1,7 @@
 #! /usr/bin/env sh
 
 # ELA:
-wget -O ../db/json/ela.json http://asn.jesandco.org/sites/default/files/ASN/manifest/D10003FC.json
+wget -O ../db/json/ccss_ela.json http://asn.jesandco.org/sites/default/files/ASN/manifest/D10003FC.json
 
 # MATH:
-wget -O ../db/json/math.json http://asn.jesandco.org/sites/default/files/ASN/manifest/D10003FB.json
+wget -O ../db/json/ccss_math.json http://asn.jesandco.org/sites/default/files/ASN/manifest/D10003FB.json
